split out node, scaleset, and pool code
