+++ src/api-service/__app__/agent_registration/__init__.py
@@ -18,8 +18,9 @@ from ..onefuzzlib.azure.queue import get_queue_sas
 from ..onefuzzlib.azure.storage import StorageType
 from ..onefuzzlib.endpoint_authorization import call_if_agent
 from ..onefuzzlib.events import get_events
-from ..onefuzzlib.pools import Node, Pool
 from ..onefuzzlib.request import not_ok, ok, parse_uri
+from ..onefuzzlib.workers.nodes import Node
+from ..onefuzzlib.workers.pools import Pool
 
 
 def create_registration_response(machine_id: UUID, pool: Pool) -> func.HttpResponse:
