+++ src/api-service/__app__/onefuzzlib/agent_events.py
@@ -25,9 +25,9 @@ from onefuzztypes.models import (
     WorkerRunningEvent,
 )
 
-from ..onefuzzlib.pools import Node, NodeTasks
 from ..onefuzzlib.task_event import TaskEvent
 from ..onefuzzlib.tasks.main import Task
+from ..onefuzzlib.workers.nodes import Node, NodeTasks
 
 MAX_OUTPUT_SIZE = 4096
