+++ src/api-service/__app__/onefuzzlib/autoscale.py
@@ -10,8 +10,10 @@ from typing import List
 from onefuzztypes.enums import NodeState, ScalesetState
 from onefuzztypes.models import AutoScaleConfig, TaskPool
 
-from .pools import Node, Pool, Scaleset
 from .tasks.main import Task
+from .workers.nodes import Node
+from .workers.pools import Pool
+from .workers.scalesets import Scaleset
 
 
 def scale_up(pool: Pool, scalesets: List[Scaleset], nodes_needed: int) -> None:
