+++ src/api-service/__app__/timer_daily/__init__.py
@@ -9,9 +9,9 @@ import azure.functions as func
 from onefuzztypes.enums import VmState
 
 from ..onefuzzlib.events import get_events
-from ..onefuzzlib.pools import Scaleset
 from ..onefuzzlib.proxy import Proxy
 from ..onefuzzlib.webhooks import WebhookMessageLog
+from ..onefuzzlib.workers.scalesets import Scaleset
 
 
 def main(mytimer: func.TimerRequest, dashboard: func.Out[str]) -> None:  # noqa: F841
